+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## Next release
 
 ### Bug fixes and improvements
-- server: fix parsing remote relationship json definition from 1.x server catalog on migration (fix #7906)
+
 - server: update pg_dump to be compatible with postgres 14 (#7676)
+- server: fix parsing remote relationship json definition from 1.x server catalog on migration (fix #7906)
+- server: Don't drop nested typed null fields in actions (fix #8237)
 
 ## v2.6.0-beta.1